Add ParticipantService.
Add ParticipantService.getMyEvents method to make a request to RR API for the events in which the logged participant is participant or enrolled.

Add ParticipantService.getMyCreatedEvents method to make a request to RR API for the events in which the logged participant is event creator.

Add ParticipantService.createEventInscription method to make a request to RR API to create an inscription to the event with the given id for the logged participant.

Add ParticipantService.completeEventInscription method to make a request to RR API to complete an inscription to the event with the given id that was opened by the logged participant.

Add ParticipantService.findParticipantSingedUpEvent method to make a request to RR API for the event details in which the logged participant is participant or enrolled.