Add system to cache the last requested event and refactor EnrolledEventComponent.
Add logic in AppComponent to be able to know when the current page has been reloaded from anywhere in the app.
Refactor ParticipantService and EventService to make the cache of the last requested event easier.
Rename ParticipantService to EventInscription Service.
Adapt app to the new cache system.
Refactor EnrolledEventComponent to show the selected event chat.
Refactor EnrolledEventComponent to be able to show the view of a created event to the current user at the same time as the event chat.
Rename EnrolledEventComponent to UserEventComponent.
Refactor myevents children routes.